SWDEV-211537, deduced conflicting types for parameter 'T'
float2 vector type change from (float x, float y) to (Scalar_accessor<T, Native_vec_, 0> x, Scalar_accessor<T, Native_vec_, 1> y).


(cherry picked from commit cf62ab6e8508b3c78765ae43834ff7eec66eaa07)

+++ src/gromacs/nbnxm/hip/nbnxm_cuda_kernel_utils.hip.h
@@ -385,7 +385,7 @@ static __forceinline__ __device__ float interpolate_coulomb_force_r(const cu_nbp
 
     float2 d01 = fetch_coulomb_force_r(nbparam, index);
 
-    return lerp(d01.x, d01.y, fraction);
+    return lerp((float)d01.x, (float)d01.y, fraction);
 }
 
 /*! Fetch C6 and C12 from the parameter table.
